removing uneeded files
